% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pois_test.R
\name{pois.test}
\alias{pois.test}
\title{Exact Poisson tests}
\usage{
pois.test(
  x,
  T = 1,
  r = 1,
  alternative = c("two.sided", "less", "greater"),
  conf.level = 0.95
)
}
\arguments{
\item{x}{A vector of length one or two representing the number of events.}

\item{T}{A vector of length one or two representing the time base for event count.}

\item{r}{A single positive number representing the hypothesized rate or rate ratio.}

\item{alternative}{A string indicating the alternative hypothesis and must be one of "two.sided", "greater" or "less".}

\item{conf.level}{A number indicating the confidence level for the returned confidence interval.}
}
\value{
A list of attributes of the exact Poisson test.
}
\description{
This function is used to perform an exact test of a simple null hypothesis about the
rate parameter in Poisson distribution, or for the ratio between two rate parameters.
}
\examples{
pois.test(100, 25)
pois.test(c(11,20),c(800, 3000))
pois.test(100, 25, alternative = "less", conf.level = 0.9)

}
